+++ code/getArticles.js
@@ -1,19 +1,20 @@
 import frontMatter from 'front-matter'
 
-var reqNew = require.context(
-	'../contenu/articles',
-	true,
-	/(2020|2021)(.)*\.md$/
-)
-const rawNew = [...reqNew.keys()].map((key) => [
+var req = require.context('../contenu/articles', true, /(20\d\d)(.)*\.md$/)
+const raw = [...req.keys()].map((key) => [
 	key.replace(/.\/(\d\d\d\d\/)?/g, '').replace('.md', ''),
-	reqNew(key).default,
+	req(key).default,
 ])
 
-export const newArticles = rawNew.map(([id, string]) => ({
+export const articles = raw.map(([id, string]) => ({
 	...frontMatter(string),
 	id,
 }))
+
+/* This could enable us to load home page articles (2020-2021) and differ the loading of the old ones 
+ * but we could need to have the date info in the parameter, which breaks the old wordpress URLs
+ * or load a list of article titles -> year to load the good chunk, but this becomes complicated
+ * if performance is a problem at this stage, we may be better with using Nextjs which does exactly that
 
 var reqOld = require.context(
 	'../contenu/articles',
@@ -30,3 +31,4 @@ export const oldArticles = rawOld.map(([id, string]) => ({
 	...frontMatter(string),
 	id,
 }))
+*/
